+++ src/Controller/XBeteiligungRestController.php
@@ -94,6 +94,14 @@ class XBeteiligungRestController extends APIController
                 throw new AccessDeniedException('Unauthorized');
             }
 
+            // Verify that the request has a valid routing key using a custom header specific to XBeteiligung.
+            // Simplest possible value is a.cockpit.a.00.01.00000000.a.00.00.00000000.a, whereas "01" needs to be
+            // changed according to the customer, mapped in XBeteiligungCustomerMappingService::FEDERAL_STATE_TO_SUBDOMAIN_MAP
+            $routingKey = $request->headers->get('X-Addon-XBeteiligung-RoutingKey','');
+            if ('' === $routingKey) {
+                throw new InvalidArgumentException('No routing key provided in X-Addon-XBeteiligung-RoutingKey header');
+            }
+
             // Get the request payload - simply use the raw XML content
             $xmlContent = $request->getContent();
 
@@ -107,7 +115,7 @@ class XBeteiligungRestController extends APIController
             ]);
 
             // Process the message using the same service that RabbitMQ would use
-            $responseObject = $xBeteiligungService->processXmlMessage($xmlContent);
+            $responseObject = $xBeteiligungService->processXmlMessage($xmlContent, true, $routingKey);
 
             // Prepare the XML response
             $xmlPayload = $responseObject->getMessageXml();